requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
python-dotenv
pypdf
pandas
openpyxl